Switch processing order in Fonduer tutorial
Changes to be committed:
      modified:   hardware/max_storage_temp_tutorial.ipynb
      modified:   hardware_image/transistor_image_tutorial.ipynb
      modified:   wiki/president_place_of_birth_tutorial.ipynb

(1) Switching the Multimodal Featurization process to ahead of training the Discriminative Model process.
(2) Fixing the tutorial sentence along with the outline in Fonduer paper.
 (1) is because Fonduer’s Multimodal Featurization process affects not the Generative Model but the Discriminative Model.
Previous version is a little bit confusing to understand the difference between Discriminative Model and Generative Model.
 In regards of (2), I guess that previous version intends collecting static process before the iterative process.
However, to understand Generative Model , Discriminative Model and Fonduer’s Multimodal Featurization,
it might be a better way to switch tutorial order and fix the tutorial sentence along with the switching.